requirements.prod.txt
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg
drf-yasg
gunicorn